+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `AngrEmulator` based on angr symbolic execution.
+- `AngrNWBTAnalysis` unused value analysis using angr.
+- `Filter` - analyses that simply listen to the hint stream.
+
+### Changed
+- `State.map()` automatically selects names for mapped objects when not
+  provided.
+
 ## [0.0.1] - 2024-02-26
 
 ### Added